backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
email-validator==2.0.0
python-dotenv==1.0.0
selenium>=4.0.0
chromedriver-autoinstaller>=0.4.0
